Refactor ReturnInsn class
